fix: product와 연관관계 매핑을 제거 및 테스트코드 일부 작성
