+++ src/main/java/com/example/i_commerce/domain/chat/entity/ChatRoom.java
@@ -39,9 +39,12 @@ public class ChatRoom extends BaseEntity {
     @Builder.Default
     private Boolean isGroupChat = false;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id")
-    private Product product;
+    //    결합도 끊어서 상품Id로만 사용하도록 개선
+//    @ManyToOne(fetch = FetchType.LAZY)
+//    @JoinColumn(name = "product_id")
+//    private Product product;
+    @Column(name = "product_id")
+    private Long productId;
 
     @Builder.Default
     @OneToMany(mappedBy = "chatRoom", cascade = CascadeType.ALL)
