+++ src/test/java/com/example/i_commerce/domain/chat/service/oldtest/ChatReportServiceTest.java
@@ -1,29 +1,25 @@
-package com.example.i_commerce.domain.chat.service;
+package com.example.i_commerce.domain.chat.service.oldtest;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
 import com.example.i_commerce.domain.chat.entity.ChatMessage;
-import com.example.i_commerce.domain.chat.entity.ChatReport;
 import com.example.i_commerce.domain.chat.entity.ChatRoom;
 import com.example.i_commerce.domain.chat.entity.enums.ChatReportReason;
-import com.example.i_commerce.domain.chat.entity.enums.ChatReportStatus;
 import com.example.i_commerce.domain.chat.repository.ChatMessageRepository;
 import com.example.i_commerce.domain.chat.repository.ChatReportRepository;
-import com.example.i_commerce.domain.member.entity.Admin;
+import com.example.i_commerce.domain.chat.service.ChatReportService;
+import com.example.i_commerce.domain.chat.service.dto.ChatReportRequest;
 import com.example.i_commerce.domain.member.entity.Member;
-import com.example.i_commerce.domain.member.entity.enums.AdminRole;
-import com.example.i_commerce.domain.member.entity.enums.AdminStatus;
 import com.example.i_commerce.domain.member.entity.enums.Gender;
 import com.example.i_commerce.domain.member.entity.enums.MemberStatus;
 import com.example.i_commerce.domain.member.entity.enums.MemberType;
-import com.example.i_commerce.domain.member.repository.AdminRepository;
 import com.example.i_commerce.domain.member.repository.MemberRepository;
 import com.example.i_commerce.global.common.response.ApiResponse;
 import com.example.i_commerce.global.security.principal.CustomUserPrincipal;
 import com.example.i_commerce.global.security.principal.CustomUserPrincipal.PrincipalType;
 import java.util.List;
 import java.util.Optional;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -35,36 +31,40 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 
 @ExtendWith(MockitoExtension.class)
-public class ChatAdminReportTest {
+public class ChatReportServiceTest {
+
     @InjectMocks
     private ChatReportService chatReportService;
 
     @Mock
     private ChatReportRepository chatReportRepository;
     @Mock
-    private AdminRepository adminRepository;
+    private MemberRepository memberRepository;
     @Mock
     private ChatMessageRepository chatMessageRepository;
 
-    private Admin admin;
+    private Member member;
     private ChatRoom chatRoom;
     private ChatMessage chatMessage;
-    private ChatReport chatReport;
 
     @BeforeEach
     public void init() {
-        CustomUserPrincipal customUserPrincipal = new CustomUserPrincipal(PrincipalType.ADMIN, 1L,
-            "admin1@naver.com", "1234", List.of());
+        CustomUserPrincipal customUserPrincipal = new CustomUserPrincipal(PrincipalType.MEMBER, 1L,
+            "test1@naver.com", "1234", List.of());
         SecurityContextHolder.getContext().setAuthentication(
             new UsernamePasswordAuthenticationToken(customUserPrincipal, null, List.of()));
 
-        admin = Admin.builder()
+        member = Member.builder()
             .id(1L)
-            .emailHash("admin1@test.com")
+            .emailHash("user1@test.com")
             .password("1234")
-            .name("어드민 유저1".getBytes())
-            .adminStatus(AdminStatus.ACTIVE)
-            .adminRole(AdminRole.MASTER)
+            .name("테스트 유저1".getBytes())
+            .sex(Gender.MALE)
+            .birthday("20050505".getBytes())
+            .phoneNumber("01011111111".getBytes())
+            .point(0)
+            .status(MemberStatus.ACTIVE)
+            .role(MemberType.CUSTOMER)
             .build();
         chatRoom = ChatRoom.builder()
             .id(1L)
@@ -77,24 +77,18 @@ public class ChatAdminReportTest {
             .memberId(2L)
             .content("테스트 메시지입니다.")
             .build();
-        chatReport = ChatReport.builder()
-            .id(1L)
-            .chatMessage(chatMessage)
-            .reporterId(1L)
-            .reportedId(2L)
-            .reason(ChatReportReason.SWEARWORD)
-            .status(ChatReportStatus.PENDING)
-            .build();
     }
+
     @Test
-    @DisplayName("관리자가 유저의 채팅의 신고를 정상적으로 처리할 수 있다.")
-    public void controlReportTest() {
-        when(chatReportRepository.findById(chatReport.getId())).thenReturn(Optional.of(chatReport));
+    @DisplayName("신고가 정상적으로 생성됩니다.")
+    public void createChatReport() {
+        ChatReportRequest request = new ChatReportRequest(1L, ChatReportReason.SWEARWORD);
+        when(memberRepository.findById(1L)).thenReturn(Optional.of(member));
+        when(chatMessageRepository.findById(1L)).thenReturn(Optional.of(chatMessage));
+        when(chatReportRepository.existsByReporterIdAndChatMessageId(1L, chatMessage.getId())).thenReturn(false);
 
-        ApiResponse<Void>response = chatReportService.controlReport(chatReport.getId(), ChatReportStatus.RESOLVED);
+        ApiResponse<Void> response = chatReportService.createChatReport(request);
 
-        Assertions.assertEquals("SUCCESS", response.code());
-        Assertions.assertEquals(ChatReportStatus.RESOLVED, chatReport.getStatus());
-        Assertions.assertTrue(chatMessage.isBlind());
+        assertEquals("SUCCESS", response.code());
     }
 }
