+++ src/test/java/com/example/i_commerce/domain/chat/service/oldtest/ChatServiceTest.java
@@ -1,4 +1,5 @@
-package com.example.i_commerce.domain.chat.service;
+
+package com.example.i_commerce.domain.chat.service.oldtest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -10,18 +11,18 @@ import com.example.i_commerce.domain.chat.exception.ChatErrorCode;
 import com.example.i_commerce.domain.chat.repository.ChatMessageRepository;
 import com.example.i_commerce.domain.chat.repository.ChatParticipantRepository;
 import com.example.i_commerce.domain.chat.repository.ChatRoomRepository;
+import com.example.i_commerce.domain.chat.service.ChatService;
+import com.example.i_commerce.domain.chat.service.dto.ChatMessageSendResponse;
 import com.example.i_commerce.domain.chat.util.ChatRoleChecker;
 import com.example.i_commerce.domain.chat.util.ChatRoomNameGenerator;
 import com.example.i_commerce.domain.member.entity.Member;
 import com.example.i_commerce.domain.member.entity.enums.Gender;
 import com.example.i_commerce.domain.member.entity.enums.MemberStatus;
 import com.example.i_commerce.domain.member.entity.enums.MemberType;
-import com.example.i_commerce.domain.member.repository.MemberRepository;
 import com.example.i_commerce.domain.product.entity.Category;
 import com.example.i_commerce.domain.product.entity.Product;
 import com.example.i_commerce.domain.product.entity.ProductOptionType;
 import com.example.i_commerce.domain.product.entity.ProductStatus;
-import com.example.i_commerce.domain.product.repository.ProductRepository;
 import com.example.i_commerce.global.common.response.ApiResponse;
 import com.example.i_commerce.global.exception.AppException;
 import com.example.i_commerce.global.security.principal.CustomUserPrincipal;
@@ -41,9 +42,10 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 
 @ExtendWith(MockitoExtension.class)
-public class ChatRoomServiceTest {
+class ChatServiceTest {
+
     @InjectMocks
-    private ChatRoomService chatRoomService;
+    private ChatService chatService;
     @Mock
     private ChatMessageRepository chatMessageRepository;
     @Mock
@@ -51,9 +53,9 @@ public class ChatRoomServiceTest {
     @Mock
     private ChatParticipantRepository chatParticipantRepository;
     @Mock
-    private MemberRepository memberRepository;
-    @Mock
-    private ProductRepository productRepository;
+    private ChatRoomNameGenerator chatRoomNameGenerator;
+    @Spy
+    private ChatRoleChecker chatRoleChecker = new ChatRoleChecker();
 
     private Member member;
     private Member member2;
@@ -64,20 +66,18 @@ public class ChatRoomServiceTest {
     private Product product;
     private ChatParticipant chatParticipant;
     private Category category;
-    @Mock
-    private ChatRoomNameGenerator chatRoomNameGenerator;
-
-    @Spy
-    private ChatRoleChecker chatRoleChecker = new ChatRoleChecker();
 
 
     @BeforeEach
     public void MemberInit() {
+/*      service 코드에서 memberId를 읽어올 때, Token의 SecurityContextHolder에서 각 값들을 빼서 사용하게 됨
+        테스트 코드에서는 직접적인 HTTP 요청이 없으므로 이 값이 존재하지 않음 그러므로 사용해주기 위해 임의의 값을 채워주어야함
+*/
         CustomUserPrincipal customUserPrincipal = new CustomUserPrincipal(PrincipalType.MEMBER, 1L,
             "test1@naver.com", "1234", List.of());
         SecurityContextHolder.getContext().setAuthentication(new UsernamePasswordAuthenticationToken(customUserPrincipal, null, List.of()));
 
-        // 카테고리 생성
+            // 카테고리 생성
         category = Category.builder()
             .id(1L)
             .name("테스트 카테고리")
@@ -155,121 +155,26 @@ public class ChatRoomServiceTest {
 
     }
 
-//  1:1 채팅방 테스트코드
-
+//    채팅내역 조회
     @Test
-    @DisplayName("1:1 채팅방이 정상적으로 생성됩니다.")
-    void successCreatePrivateChatRoom() {
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(memberRepository.findById(otherMember.getId())).thenReturn(Optional.of(otherMember));
-        when(chatParticipantRepository.findExistingPrivateRoom
-            (member.getId(), otherMember.getId())).thenReturn(Optional.empty());
-
-        ApiResponse<Long> response = chatRoomService.getOrCreatePrivateRoom(otherMember.getId());
-
-        assertEquals("SUCCESS", response.code());
-
-    }
-
-    @Test
-    @DisplayName("1:1 채팅방이 이미 존재하면 예외를 터트립니다.")
-    void existingChatRoom() {
-
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(memberRepository.findById(otherMember.getId())).thenReturn(Optional.of(otherMember));
-        when(chatParticipantRepository.findExistingPrivateRoom(member.getId(),
-            otherMember.getId())).thenReturn(Optional.of(singlechatRoom));
-
-        AppException exception = assertThrows(AppException.class,
-            () -> chatRoomService.getOrCreatePrivateRoom(otherMember.getId()));
-
-        Assertions.assertThat(exception.getErrorCode())
-            .isEqualTo(ChatErrorCode.CHAT_ROOM_ALREADY_EXISTS);
-    }
-
-    @Test
-    @DisplayName("1:1 채팅 생성시 서로의 권한이 같으면 에러를 터트립니다.")
-    void cannotSameRole(){
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when((memberRepository.findById(member2.getId()))).thenReturn(Optional.of(member2));
-
+    @DisplayName("채팅에 참여중이 아닌 사람이 조회를 시도하면 오류를 발생시킵니다.")
+    void getChatHistory_NoMember(){
+        when(chatRoomRepository.findById(singlechatRoom.getId())).thenReturn(Optional.of(singlechatRoom));
+        when(chatParticipantRepository.findByChatRoom(singlechatRoom)).thenReturn(List.of());
         AppException exception = assertThrows(AppException.class, () ->
-            chatRoomService.getOrCreatePrivateRoom(member2.getId())); //
-
-        Assertions.assertThat(exception.getErrorCode()).isEqualTo(ChatErrorCode.CANNOT_CHAT_SAME_ROLE);
+            chatService.getChatHistory(singlechatRoom.getId()));
+        Assertions.assertThat(exception.getErrorCode()).isEqualTo(ChatErrorCode.NOT_A_ROOM_MEMBER);
     }
-
-    //    그룹채팅 테스트코드
     @Test
-    @DisplayName("그룹 채팅방이 이미 존재하면 예외를 터트립니다.")
-    void createGroupChatRoom() {
+    @DisplayName("채팅 내역을 정상적으로 불러옵니다")
+    void getChatHistorySuccess(){
+        when(chatRoomRepository.findById(singlechatRoom.getId())).thenReturn(Optional.of(singlechatRoom));
+        when(chatParticipantRepository.findByChatRoom(singlechatRoom)).thenReturn(List.of(chatParticipant));
+        when(chatMessageRepository.findByChatRoomOrderByCreatedAtAsc(singlechatRoom)).thenReturn(List.of());
 
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(productRepository.findById(product.getId())).thenReturn(Optional.of(product));
-        when(chatRoomRepository.existsByProductIdAndIsGroupChat(product.getId(), true)).thenReturn(
-            true);
-
-        AppException exception = assertThrows(AppException.class, () ->
-            chatRoomService.createGroupRoom(product.getId()));
-
-        Assertions.assertThat(exception.getErrorCode())
-            .isEqualTo(ChatErrorCode.CHAT_ROOM_ALREADY_EXISTS);
-    }
-
-    @Test
-    @DisplayName("그룹 채팅방이 없으면 새로 생성합니다.")
-    void createGroupRoomSuccess() {
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(productRepository.findById(product.getId())).thenReturn(Optional.of(product));
-        when(chatRoomRepository.existsByProductIdAndIsGroupChat(product.getId(), true)).thenReturn(
-            false);
-
-        ApiResponse<Long> response = chatRoomService.createGroupRoom(product.getId());
-
+        ApiResponse<List<ChatMessageSendResponse>> response = chatService.getChatHistory(singlechatRoom.getId());
         assertEquals("SUCCESS", response.code());
-    }
-
-
-    @Test
-    @DisplayName("그룹 채팅방에 참여합니다.")
-    void joinGroupChatRoom() {
-        //when
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(chatRoomRepository.findById(groupchatRoom.getId())).thenReturn(
-            Optional.of(groupchatRoom));
-
-        ApiResponse<Void> response = chatRoomService.joinGroupRoom(groupchatRoom.getId());
-
-        assertEquals("SUCCESS", response.code());
-    }
-
-//        @Test
-//    @DisplayName("그룹 채팅방에서 퇴장합니다.")
-//    void leaveGroupRoom() {
-//        when(chatRoomRepository.findById(groupchatRoom.getId())).thenReturn(
-//            Optional.of(groupchatRoom));
-//        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-//        when(chatParticipantRepository.findByChatRoomAndMember(groupchatRoom, member))
-//            .thenReturn(Optional.of(chatParticipant));
-//
-//        ApiResponse<Void> response = chatService.leaveGroupRoom(groupchatRoom.getId(),
-//            member.getId());
-//
-//        assertEquals("SUCCESS", response.code());
-//
-//    }
-    @Test
-    @DisplayName("그룹 채팅방에서 퇴장합니다.")
-    void leaveGroupRoom() {
-        when(chatRoomRepository.findById(groupchatRoom.getId())).thenReturn(
-            Optional.of(groupchatRoom));
-        when(memberRepository.findById(member.getId())).thenReturn(Optional.of(member));
-        when(chatParticipantRepository.findByChatRoomAndMemberId(groupchatRoom, member.getId()))
-            .thenReturn(Optional.of(chatParticipant));
-
-        ApiResponse<Void> response = chatRoomService.leaveGroupRoom(groupchatRoom.getId());
-
-        assertEquals("SUCCESS", response.code());
-
     }
 }
+
+
